refactor: consolidate duplicated styles into common.styles.ts
- Rename shared.styles.ts → common.styles.ts for clarity
- Extract all duplicated patterns into common styles:
  - container (width: 100%)
  - heading (section h2 with margin)
  - subheading (uppercase, heading font, standard margin)
  - title (h3 with no margin)
  - textPrimary (primary color utility)
- Simplify component style files to re-export common styles
- About, Education, Experience, Skills now only export aliased common styles
- Projects only has unique styles (projectList, projectItem) + common re-exports
- Footer and Page re-export from common.styles instead of duplicating
- Update 404 page to use common.styles

Result: Eliminated ~100 lines of duplicated code across 6+ files.
Bundle size reduced: CSS 167.72 KB (from 167.78 KB), JS 316.12 KB (from 316.82 KB)
All 49 tests pass.
