+++ src/resume/skills.styles.ts
@@ -1,20 +1,2 @@
-import { css } from '../../styled-system/css';
-
-// Container for the entire skills section
-export const container = css({
-  width: '100%',
-});
-
-// Section heading styling
-export const heading = css({
-  marginBottom: 'headingMargin',
-});
-
-// Skill category title
-export const skillTitle = css({
-  marginBottom: 'subheadingMargin',
-  textTransform: 'uppercase',
-  fontWeight: 500,
-  fontSize: '1.5rem',
-  fontFamily: 'heading',
-});
+// Re-export common styles
+export { container, heading, subheading as skillTitle } from './common.styles';
